separate chart into another file
